+++ src/FunderMaps.Core/Services/NullEmailService.cs
@@ -14,18 +14,12 @@ namespace FunderMaps.Core.Services
         ///     Send email message.
         /// </summary>
         /// <param name="emailMessage">Message to send.</param>
-        public Task SendAsync(EmailMessage emailMessage)
-        {
-            return Task.CompletedTask;
-        }
+        public Task SendAsync(EmailMessage emailMessage) => Task.CompletedTask;
 
         /// <summary>
-        ///     Test the email service backend.
+        ///     Test the batch service backend.
         /// </summary>
-        public Task TestService()
-        {
-            return Task.CompletedTask;
-        }
+        public Task HealthCheck() => Task.CompletedTask;
     }
 }
 #pragma warning restore CA1812 // Internal class is never instantiated
